requirements.txt
groq
streamlit
python-dotenv
Pillow
matplotlib
numpy